Add html-webpack-plugin, reorganize sources
The html-webpack-plugin takes care of applying a template `index.html`
file to the build directory, appropriately referencing (linking) required
assets.

To accommodate the plugin, the structure of the `source` directory and its
subdirectories has been changed. The `index.html` and `index.js` files are
now located in the same directory, with the application sources and
paraphernalia placed in the `source/app` directory.

The Webpack debug configuration reflects the changes made to the project
structure.