Begin work on Login page
The Login page presents a form that requires a username, password, and
email input. These should be saved by the application's state and used
throughout the application, even on other pages.

The source directory structure has (again) changed to conform with the
application's design. Every page is located in its own directory, and
every page has its own container module to which application routes refer
to.