Overhaul app structure
Component (and container) modules are renamed to conform to standards
observable in OOP languages, because the OOP approach of ES2015 is used in
application source files.

Page modules are created by merging together the container and component
modules of a page.

Application routing now expects a user ID parameter. Data should be
fetched based on user ID. A placeholder UserData module is included for
reference.