Add stylesheets, update component views
SCSS is used to style the application. Style sheets are located in
`app/assets/styles/`, and are integrated in the application during the
packaging process.

Component and page views (layouts) are changed to reflect the stylesheet
rules.

Two new components are defined. `ImageEdit` is a stub for a component that
should allow changing the displayed image. `UserInfo` is a stateless
component that displays user's full name and email.